staples project

+++ src/main/java/util/TestBase.java
@@ -7,6 +7,11 @@ import java.util.concurrent.TimeUnit;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import io.github.bonigarcia.wdm.WebDriverManager;
 
 public class TestBase {
 
@@ -32,13 +37,28 @@ public class TestBase {
 
 			System.setProperty("webdriver.chrome.driver","Driver/chromedriver.exe");
 			driver = new ChromeDriver();
+		}else if (browserName.equalsIgnoreCase("firefox")) {
+
+				System.setProperty("webdriver.gecko.driver","Driver/geckodriver.exe");
+				driver = new FirefoxDriver();
+			}else
+			{
+				
+				System.out.println("No Matching browser found for  " + prop.getProperty("browserName"));
+			}
+		
+//			 WebDriverManager.chromedriver().setup();
+//			 WebDriver driver = new ChromeDriver();
 			driver.manage().window().maximize();
 			driver.manage().deleteAllCookies();
 			driver.manage().timeouts().pageLoadTimeout(TestUtil.PAGE_LOAD_TIMEOUT, TimeUnit.SECONDS);
 			driver.manage().timeouts().implicitlyWait(TestUtil.IMPLICIT_WAIT, TimeUnit.SECONDS);
 			driver.get(prop.getProperty("url"));
 		}
+	
+	
+	
 
 	}
 
-}
+
